Some unreleased code reverted
  -- improvement made for internal id collision is reverted as the
JavaScript doesn't support 64 bit int

+++ wffweb/src/main/java/com/webfirmframework/wffweb/tag/html/model/AbstractHtml5SharedObject.java
@@ -23,7 +23,7 @@ import java.util.Set;
 import java.util.WeakHashMap;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.atomic.AtomicBoolean;
-import java.util.concurrent.atomic.AtomicLong;
+import java.util.concurrent.atomic.AtomicInteger;
 import java.util.concurrent.atomic.AtomicReference;
 import java.util.concurrent.locks.ReentrantReadWriteLock;
 
@@ -83,10 +83,11 @@ public class AbstractHtml5SharedObject implements Serializable {
     private PushQueue pushQueue;
 
     /**
-     * no need to make it volatile declared as long to avoid maximum id
-     * collision
+     * no need to make it volatile. cannot declare as long to avoid maximum id
+     * collision because JavaScript doesn't support 64 bit int as of now. If it
+     * is modified DataWffIdUtil.getDataWffIdBytes should also be modified
      */
-    private final AtomicLong dataWffId = new AtomicLong(-1);
+    private final AtomicInteger dataWffId = new AtomicInteger(-1);
 
     private final AtomicBoolean dataWffIdSecondCycle = new AtomicBoolean(false);
 
@@ -139,14 +140,14 @@ public class AbstractHtml5SharedObject implements Serializable {
 
         while (tagByWffId.size() < Integer.MAX_VALUE) {
 
-            final long incrementedDataWffId = dataWffId.incrementAndGet();
+            final int incrementedDataWffId = dataWffId.incrementAndGet();
 
             if (incrementedDataWffId < 0 || dataWffIdSecondCycle.get()) {
 
                 dataWffIdSecondCycle.compareAndSet(false, true);
 
-                long newDataWffId = incrementedDataWffId < 0
-                        ? ((incrementedDataWffId - Long.MAX_VALUE) - 1)
+                int newDataWffId = incrementedDataWffId < 0
+                        ? ((incrementedDataWffId - Integer.MAX_VALUE) - 1)
                         : incrementedDataWffId;
 
                 String id = "S" + (newDataWffId);
@@ -154,7 +155,7 @@ public class AbstractHtml5SharedObject implements Serializable {
                 while (tagByWffId.containsKey(id)) {
                     newDataWffId++;
                     if (newDataWffId < 0) {
-                        newDataWffId = (newDataWffId - Long.MAX_VALUE) - 1;
+                        newDataWffId = (newDataWffId - Integer.MAX_VALUE) - 1;
                     }
                     id = "S" + newDataWffId;
                 }
@@ -191,7 +192,7 @@ public class AbstractHtml5SharedObject implements Serializable {
         return lock;
     }
 
-    public long getLastDataWffId(final Object accessObject) {
+    public int getLastDataWffId(final Object accessObject) {
 
         if (accessObject == null || !((SecurityClassConstants.ABSTRACT_HTML
                 .equals(accessObject.getClass().getName())))) {
